rename AbstractMongoCRUDTestController to LightblueMongoTestHarness
